#280: Moved NOT_A_VALUE from CoreValue back to Value.
