+++ core/src/main/java/io/parsingdata/metal/expression/comparison/ComparisonExpression.java
@@ -19,7 +19,7 @@ package io.parsingdata.metal.expression.comparison;
 import static io.parsingdata.metal.Trampoline.complete;
 import static io.parsingdata.metal.Trampoline.intermediate;
 import static io.parsingdata.metal.Util.checkNotNull;
-import static io.parsingdata.metal.expression.value.CoreValue.NOT_A_VALUE;
+import static io.parsingdata.metal.expression.value.Value.NOT_A_VALUE;
 
 import java.util.Objects;
