+++ core/src/main/java/io/parsingdata/metal/token/RepN.java
@@ -18,7 +18,7 @@ package io.parsingdata.metal.token;
 
 import static io.parsingdata.metal.Util.checkNotNull;
 import static io.parsingdata.metal.Util.failure;
-import static io.parsingdata.metal.expression.value.CoreValue.NOT_A_VALUE;
+import static io.parsingdata.metal.expression.value.Value.NOT_A_VALUE;
 
 import java.math.BigInteger;
 import java.util.Objects;
