+++ core/src/test/java/io/parsingdata/metal/expression/value/FoldEdgeCaseTest.java
@@ -34,7 +34,7 @@ import static io.parsingdata.metal.Shorthand.seq;
 import static io.parsingdata.metal.data.Slice.createFromBytes;
 import static io.parsingdata.metal.encoding.Encoding.DEFAULT_ENCODING;
 import static io.parsingdata.metal.expression.value.BytesTest.EMPTY_PARSE_STATE;
-import static io.parsingdata.metal.expression.value.CoreValue.NOT_A_VALUE;
+import static io.parsingdata.metal.expression.value.Value.NOT_A_VALUE;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static io.parsingdata.metal.util.EnvironmentFactory.env;
 import static io.parsingdata.metal.util.ParseStateFactory.stream;
